[Fix] OpenGL rendering
The Rendering was shit before. Now it just uses legacy glBegin and glEnd
for everything and it works good and I am happy. No ShaderProgramme,
VAO, VBO, EBO or something else needed earlier. It just gets my work
done. And is easy to understand.

Before it drew a solid avg colour of image, now it draws the image :)

Note to myself: Don't try OpenGL-v2.1 with OpenGL-v3 tutorials or
concepts.